refactor(event-stream): extract I/O thread to connection.rs
Pure code motion for #6235. Move the event-stream I/O thread out of
mod.rs into a dedicated connection.rs submodule: io_thread_main (the
reconnect loop), try_connect (socket retry), replay_buffered (re-send of
un-ACKed frames, parking the ordered replay-gap FullResync barrier
#5267), write_all_backpressured (the stop-aware nonblocking backpressured
writer #2877 shared with the drain path), and run_connected_loop (the
steady-state drain / write / read-control / keepalive cycle #2883).

Moved verbatim (no logic change). io_thread_main, replay_buffered,
write_all_backpressured, and run_connected_loop are widened to pub(super)
so mod.rs (EventStreamSender::new), control.rs (handle_drain_request's
shared writer), and the backpressure / replay_budget / drain test suites
resolve them; try_connect stays private (only io_thread_main calls it).
The std imports these functions used (VecDeque / io / UnixStream /
TryRecvError) now resolve for the submodules via `use super::*` and are
annotated #[allow(unused_imports)] in the facade. cargo build green.
